requirements.txt
paddlehub==1.8.2
paddlepaddle==1.8.5
Pillow==7.2.0
